Add scripts for ECM, Trophy, and cargo paradrops
The ECM script causes a randomly selected proportion of missiles
that get within a certain distance of a vehicle to veer off as
if decoyed by a repeater jammer. The Trophy script causes a
certain group of projectile types to be destroyed after getting
within a certain distance of a vehicle, but with a time delay
between the destruction of each projectile, allowing it to be
overwhelmed with volley fire. The cargo paradrop script is self
explanatory, adding an alternative to using the RHS mod's cargo
paradrop waypoint.